+++ frappe/core/doctype/system_settings/system_settings.json
@@ -37,7 +37,7 @@
   "allow_login_using_mobile_number",
   "allow_login_using_user_name",
   "allow_error_traceback",
-  "remove_exif_tags",
+  "strip_exif_metadata_from_uploaded_images",
   "password_settings",
   "logout_on_password_reset",
   "force_user_to_reset_password",
@@ -464,15 +464,15 @@
   },
   {
    "default": "1",
-   "fieldname": "remove_exif_tags",
+   "fieldname": "strip_exif_metadata_from_uploaded_images",
    "fieldtype": "Check",
-   "label": "Remove EXIF tags from uploaded images"
+   "label": "Strip EXIF tags from uploaded images"
   }
  ],
  "icon": "fa fa-cog",
  "issingle": 1,
  "links": [],
- "modified": "2020-11-30 17:04:06.785282",
+ "modified": "2020-11-30 18:52:22.161391",
  "modified_by": "Administrator",
  "module": "Core",
  "name": "System Settings",
